+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.1.0] - 2018-06-28
 
 ### Added
 
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Restructured source files
 * Don't allow the FileSystem to be copied
 * Mount filesystem externally
+* Use <> includes
+* Use Sematic Versioning correctly
 
 ## [0.0.2] - 2018-02-11